src/requirements.txt
alembic==1.8.1
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
